Extract Firestore client wrapper to its own package.
