+++ sync/firestore/firestore_test.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 	"testing"
 	"time"
-
+	fsClient "github.com/wptide/pkg/wrapper/firestore"
 	"github.com/wptide/pkg/wporg"
 )
 
@@ -55,14 +55,14 @@ func (m mockClient) Authenticated() bool {
 	return true
 }
 
-func (m mockClient) QueryItems(collection string, conditions []Condition, ordering []Order, limit int, updateFunc UpdateFunc) ([]interface{}, error) {
+func (m mockClient) QueryItems(collection string, conditions []fsClient.Condition, ordering []fsClient.Order, limit int, updateFunc fsClient.UpdateFunc) ([]interface{}, error) {
 	return nil, nil
 }
 
 func TestFirestoreSync_UpdateCheck(t *testing.T) {
 	type fields struct {
 		ctx      context.Context
-		client   ClientInterface
+		client   fsClient.ClientInterface
 		rootPath string
 	}
 	type args struct {
@@ -130,7 +130,7 @@ func TestFirestoreSync_UpdateCheck(t *testing.T) {
 func TestFirestoreSync_RecordUpdate(t *testing.T) {
 	type fields struct {
 		ctx      context.Context
-		client   ClientInterface
+		client   fsClient.ClientInterface
 		rootPath string
 	}
 	type args struct {
@@ -177,7 +177,7 @@ func TestFirestoreSync_RecordUpdate(t *testing.T) {
 func TestFirestoreSync_SetSyncTime(t *testing.T) {
 	type fields struct {
 		ctx      context.Context
-		client   ClientInterface
+		client   fsClient.ClientInterface
 		rootPath string
 	}
 	type args struct {
@@ -229,7 +229,7 @@ func TestFirestoreSync_GetSyncTime(t *testing.T) {
 
 	type fields struct {
 		ctx      context.Context
-		client   ClientInterface
+		client   fsClient.ClientInterface
 		rootPath string
 	}
 	type args struct {
@@ -417,7 +417,7 @@ func TestNewWithClient(t *testing.T) {
 		ctx         context.Context
 		projectId   string
 		rootDocPath string
-		client      ClientInterface
+		client      fsClient.ClientInterface
 	}
 	tests := []struct {
 		name string
