+++ wrapper/firestore/client.go
